refactor(frontend): update imports for renamed modules
