+++ server/zanata-frontend/src/frontend/app/editor/components/LanguagesDropdown.js
@@ -1,5 +1,5 @@
 import { values } from 'lodash'
-import { encode } from '../utils/doc-id'
+import { encode } from '../utils/doc-id-util'
 import Dropdown from './Dropdown'
 import { Icon, Row } from 'zanata-ui'
 import React, { PropTypes } from 'react'
