+++ server/zanata-frontend/src/frontend/app/editor/components/TransUnitTranslationFooter.js
@@ -24,8 +24,9 @@ import cx from 'classnames'
 import Button from './Button'
 import SplitDropdown from './SplitDropdown'
 import { Icon, Row } from 'zanata-ui'
-import { defaultSaveStatus, nonDefaultValidSaveStatuses } from '../utils/status'
-import { hasTranslationChanged } from '../utils/phrase'
+import { defaultSaveStatus, nonDefaultValidSaveStatuses }
+  from '../utils/status-util'
+import { hasTranslationChanged } from '../utils/phrase-util'
 
 /**
  * Footer for translation with save buttons and other action widgets.
