+++ server/zanata-frontend/src/frontend/app/editor/containers/KeyShortcutCheatSheet/index.js
@@ -4,8 +4,8 @@ import KeyCombinations from '../../components/KeyCombinations'
 import { chain, each, map } from 'lodash'
 import { connect } from 'react-redux'
 import React, { PropTypes } from 'react'
-import { SHORTCUTS } from '../../actions/editorShortcuts'
-import { toggleKeyboardShortcutsModal } from '../../actions/headerActions'
+import { SHORTCUTS } from '../../actions/key-shortcuts-actions'
+import { toggleKeyboardShortcutsModal } from '../../actions/header-actions'
 
 /**
  * Modal showing a summary of the available key shortcuts.
