+++ server/zanata-frontend/src/frontend/app/editor/containers/Root/index.js
@@ -7,8 +7,8 @@ import KeyShortcutCheatSheet from '../KeyShortcutCheatSheet'
 import KeyShortcutDispatcher from '../KeyShortcutDispatcher'
 import SuggestionsPanel from '../SuggestionsPanel'
 import { setSidebarVisibility } from '../../actions'
-import { fetchUiLocales } from '../../actions/headerActions'
-import { saveSuggestionPanelHeight } from '../../actions/suggestions'
+import { fetchUiLocales } from '../../actions/header-actions'
+import { saveSuggestionPanelHeight } from '../../actions/suggestions-actions'
 import SplitPane from 'react-split-pane'
 import { Icons } from 'zanata-ui'
 import Sidebar from '../Sidebar'
