+++ server/zanata-frontend/src/frontend/app/editor/reducers/header-data-reducer.js
@@ -6,7 +6,7 @@ import {
   HEADER_DATA_FETCHED,
   LOCALE_SELECTED,
   STATS_FETCHED
-} from '../actions/headerActions'
+} from '../actions/header-actions'
 import updateObject from 'react-addons-update'
 import {prepareLocales, prepareStats, prepareDocs} from '../utils/Util'
 import { dashboardUrl, serviceUrl } from '../api'
