+++ server/zanata-frontend/src/frontend/app/editor/reducers/phrase-reducer.js
@@ -2,8 +2,8 @@ import updateObject from 'react-addons-update'
 import {
   CLAMP_PAGE,
   UPDATE_PAGE
-} from '../actions/controlsHeaderActions'
-import { COPY_GLOSSARY_TERM } from '../actions/glossary'
+} from '../actions/controls-header-actions'
+import { COPY_GLOSSARY_TERM } from '../actions/glossary-actions'
 import {
   CANCEL_EDIT,
   COPY_FROM_ALIGNED_SOURCE,
@@ -21,16 +21,16 @@ import {
   SELECT_PHRASE_SPECIFIC_PLURAL,
   TRANSLATION_TEXT_INPUT_CHANGED,
   UNDO_EDIT
-} from '../actions/phrases'
-import { COPY_SUGGESTION } from '../actions/suggestions'
+} from '../actions/phrases-actions'
+import { COPY_SUGGESTION } from '../actions/suggestions-actions'
 import {
   calculateMaxPageIndex,
   calculateMaxPageIndexFromState,
   getFilteredPhrasesFromState
 } from '../utils/filter-paging-util'
 import { replaceRange } from '../utils/string-utils'
-import { SET_SAVE_AS_MODE } from '../actions/editorShortcuts'
-import { MOVE_NEXT, MOVE_PREVIOUS } from '../actions/phraseNavigation'
+import { SET_SAVE_AS_MODE } from '../actions/key-shortcuts-actions'
+import { MOVE_NEXT, MOVE_PREVIOUS } from '../actions/phrase-navigation-actions'
 
 // TODO use lodash when upgraded
 // clamps a number within the inclusive lower and upper bounds
